Added a few rules (experimenting with PyTest)
