--- tests/rules/test_rule.py
@@ -1,9 +1,0 @@
-import pytest
-
-from gmail_rules.rules.rule import Rule
-
-class test_rule:
-
-    def test_rule_definition(self):
-        new_rule = Rule()
-        assert isinstance(new_rule, Rule)
